feat: nice navbar, change name, and remove unused
Update details
- remove unused css file
- change file name for naming convention
- implement nice navbar

Update: landing page
